Move the implementation to KafkaApis
