--- core/src/main/scala/kafka/server/validator/AuthorizeFetchPartitionValidator.scala
@@ -1,45 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package kafka.server.validator
-
-import kafka.network.RequestChannel
-import kafka.security.auth.Authorizer
-import kafka.security.auth.Read
-import kafka.security.auth.Resource
-import kafka.security.auth.Topic
-import org.apache.kafka.common.protocol.Errors
-import org.apache.kafka.common.resource.PatternType.LITERAL
-
-final class AuthorizeFetchPartitionValidator(authorizer: Option[Authorizer]) extends FetchPartitionValidator {
-  import AuthorizeFetchRequestValidator.authorize
-
-  override def validate(request: RequestChannel.Request, partition: ValidElem): Option[ErrorElem] = {
-    val (topic, data) = partition
-    if (!authorize(authorizer, request.session, Read, Resource(Topic, topic.topic, LITERAL))) {
-      Some(topic -> errorResponse(Errors.TOPIC_AUTHORIZATION_FAILED))
-    } else {
-      None
-    }
-  }
-}
-
-final object AuthorizeFetchPartitionValidator {
-  def apply(authorizer: Option[Authorizer]): AuthorizeFetchPartitionValidator = {
-    new AuthorizeFetchPartitionValidator(authorizer)
-  }
-}
